Multiple features for multiple events
Duplicates features in case of multiple associated time events. I'm
worried this is somewhat "spaghetti" code, but instead of having a
separate event model, we're duplicating features to add different "when"
intervals. The point of this change is to align with ongoing standards
work led by Sean Gillies and the point he made here:
https://github.com/geojson/geojson-ld/issues/18#issuecomment-44095414