Preparing more optimized queries
A few stored procedures and some more joins will help reduce the number
of queries used to generate an view of an item. This is good since we
seem to have some severe speed issues right now. By putting more work in
the database, we can probably cut the number of queries to at least in
half of the current use.