Improved media file importing
Less likely to 500 error. Catches duplicate key record insert problems.
Adds a short pause before saving a record so remote server won't deny
the HEAD request. Saves corrected extension capitalization in the Import
Cell object model so if an import needs to be re-run (as is often the
case), we don't have to look for the correct extension capitalization
all over again.