Better metadata management of export tables
Since we need to be able to edit export table metadata, it makes sense
to use more general Open Context modes / code, rather than reinvent the
wheel. However, export table metadata needs to be somewhat different, in
that it needs to record counts of object items. Therefore, this change
includes an extension to the link_annoations model so that we can add
extra information to the assertion. That extra information is in the
form of a JSON encoded dict object.

There's a method that we will include temporarily to migrate the
existing table metadata to the link_annotations table.